Updates Readme

+++ LocalMarket/LocalMarket/LocalMarket/models.py
@@ -6,7 +6,7 @@ class Item(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     name = models.CharField(max_length=255)
     category = models.CharField(max_length=255)
-    price = models.PositiveIntegerField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
 
     def __str__(self):
         return self.name
